requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
qdrant-client>=1.11.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0

# Optional: Keep OpenAI for backward compatibility
# openai==1.10.0
